Add a protocol file and remove protocol classes from 2pcf and cls likelihood computation. change xi_plus, xi_minus vector names
