+++ cloelike/EuclidLikelihood_protocol.py
@@ -1,0 +1,65 @@
+import numpy as np
+from typing import Protocol, runtime_checkable
+from cloelib.cosmology.cosmology import Background, Perturbations
+
+
+@runtime_checkable
+class PhotoLikelihoodProtocol(Protocol):
+    """
+    Protocol for photo-z likelihood classes.
+
+    This protocol defines the required interface for photometric likelihood implementations,
+    specifying initialization, required attributes, and methods for computing data vectors,
+    theory vectors, covariance matrices, and log-likelihoods.
+
+    Attributes:
+        data (dict): Observational data dictionary.
+        settings (dict): Configuration settings dictionary.
+        Background (Background): Cosmological background instance.
+        LinPerturbations (Perturbations): Linear perturbations instance.
+        NonLinPerturbations (Perturbations): Non-linear perturbations instance.
+        derived (dict): Dictionary for derived quantities.
+        mode (str): Mode of operation (e.g., "coupled").
+
+    Methods:
+        __init__(data, settings, Background, LinPerturbations, NonLinPerturbations, mode):
+            Initializes the likelihood protocol.
+        get_data_vector_full() -> np.ndarray:
+            Returns the full data vector.
+        get_data_vector_masked() -> np.ndarray:
+            Returns the masked data vector.
+        get_theory_vector_full(parameters: dict) -> np.ndarray:
+            Returns the full theory vector for given parameters.
+        get_theory_vector_masked(parameters: dict) -> np.ndarray:
+            Returns the masked theory vector for given parameters.
+        get_covariance_matrix_full() -> np.ndarray:
+            Returns the full covariance matrix.
+        get_covariance_matrix_masked_inv() -> np.ndarray:
+            Returns the inverse of the masked covariance matrix.
+        loglike(parameters: dict) -> float:
+            Computes the log-likelihood for the given parameters.
+    """
+
+    def __init__(
+        self,
+        data: dict,
+        settings: dict,
+        Background: Background,
+        LinPerturbations: Perturbations,
+        NonLinPerturbations: Perturbations,
+    ) -> None: ...
+
+    data: dict
+    settings: dict
+    Background: Background
+    LinPerturbations: Perturbations
+    NonLinPerturbations: Perturbations
+    derived: dict
+
+    def get_data_vector_full(self) -> np.ndarray: ...
+    def get_data_vector_masked(self) -> np.ndarray: ...
+    def get_theory_vector_full(self, parameters: dict) -> np.ndarray: ...
+    def get_theory_vector_masked(self, parameters: dict) -> np.ndarray: ...
+    def get_covariance_matrix_full(self) -> np.ndarray: ...
+    def get_covariance_matrix_masked_inv(self) -> np.ndarray: ...
+    def loglike(self, parameters: dict) -> float: ...
